refactor: Move `language.pyi` stub generation logic into `build.rs` and add the generated file to `.gitignore`.
