+++ api/python/slint/.gitignore
@@ -1,3 +1,4 @@
 slint/*.so
+slint/language.pyi
 uv.lock
 docs
